requirements.txt
pymongo[srv]==3.11.3
Flask==2.0.1
gunicorn==20.0.4
python-decouple==3.4
Flask-RESTful==0.3.9
celery==4.4.0
redis==3.3.7

oauth2==1.9.0.post1
urllib3==1.26.4
